requirements.txt
absl-py==1.4.0
black==23.3.0
brotlipy==0.7.0
cachetools==5.3.1
certifi @ file:///C:/b/abs_4a0polqwty/croot/certifi_1683875377622/work/certifi
cffi @ file:///C:/b/abs_49n3v2hyhr/croot/cffi_1670423218144/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.3
colorama==0.4.6
cryptography @ file:///C:/b/abs_8ecplyc3n2/croot/cryptography_1677533105000/work
filelock @ file:///C:/b/abs_c7yrhs9uz2/croot/filelock_1672387617533/work
fsspec==2023.5.0
google-auth==2.19.1
google-auth-oauthlib==1.0.0
grpcio==1.54.2
idna @ file:///C:/b/abs_bdhbebrioa/croot/idna_1666125572046/work
Jinja2 @ file:///C:/b/abs_7cdis66kl9/croot/jinja2_1666908141852/work
Markdown==3.4.3
MarkupSafe @ file:///C:/ci/markupsafe_1654508036328/work
mkl-fft==1.3.6
mkl-random @ file:///C:/Users/dev-admin/mkl/mkl_random_1682977971003/work
mkl-service==2.4.0
mpmath==1.2.1
mypy-extensions==1.0.0
networkx @ file:///C:/b/abs_b935xy_9g6/croot/networkx_1678964342510/work
numpy @ file:///C:/Users/dev-admin/mkl/numpy_and_numpy_base_1682982345978/work
oauthlib==3.2.2
opencv-python==4.7.0.72
packaging==23.1
pandas==2.0.2
pathspec==0.11.1
Pillow==9.4.0
platformdirs==3.5.1
protobuf==4.23.2
psutil==5.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pyOpenSSL @ file:///C:/b/abs_552w85x1jz/croot/pyopenssl_1677607703691/work
pyre-extensions==0.0.23
PySocks @ file:///C:/ci_310/pysocks_1642089375450/work
python-dateutil==2.8.2
pytz==2023.3
requests @ file:///C:/b/abs_41owkd5ymz/croot/requests_1682607524657/work
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sympy @ file:///C:/b/abs_95fbf1z7n6/croot/sympy_1668202411612/work
tensorboard==2.13.0
tensorboard-data-server==0.7.0
tomli==2.0.1
torch==2.0.0
torcheval==0.0.6
torchtnt==0.1.0
torchvision==0.15.0
tqdm==4.65.0
typing-inspect==0.9.0
typing_extensions @ file:///C:/b/abs_a1bb332wcs/croot/typing_extensions_1681939523095/work
tzdata==2023.3
urllib3 @ file:///C:/b/abs_3ce53vrdcr/croot/urllib3_1680254693505/work
Werkzeug==2.3.4
win-inet-pton @ file:///C:/ci_310/win_inet_pton_1642658466512/work
xformers==0.0.18
